Rename LikedArticle to ArticleListItem
